add skeleton for recursive loop fusion function as-04
add fuseLevelNLoops, with basic iteration over the vector at current level (for now it only prints the loop pairs, and for each one deletes the second from the vector so we get to convergence at some point). still missing the recursive call on each loop nest from each loop in the vector.